refactor: 패키지 구조 변경 및 controller, domain, view 클래스 분류
