+++ src/main/java/race/RaceMain.java
@@ -1,0 +1,11 @@
+package race;
+
+import race.controller.RaceController;
+
+public class RaceMain {
+
+    public static void main(String[] args) {
+        RaceController race = new RaceController();
+        race.run();
+    }
+}
