+++ src/main/java/race/domain/Cars.java
@@ -1,4 +1,4 @@
-package race;
+package race.domain;
 
 import java.util.ArrayList;
 import java.util.List;
