+++ src/test/java/race/CarsTest.java
@@ -4,6 +4,8 @@ import org.assertj.core.api.SoftAssertions;
 import org.junit.jupiter.api.Test;
 import org.junit.jupiter.params.ParameterizedTest;
 import org.junit.jupiter.params.provider.ValueSource;
+import race.domain.Car;
+import race.domain.Cars;
 
 import java.util.List;
 import java.util.Map;
